Generating valid Moves
The problem was the way i removed moves from the list of all moves messed with the order in which the for loop went through the list. It is never a good idea to directly loop through the list that you remove elemnts from. I went by reverse index now.